source/requirements.txt
altgraph==0.17.3
defusedxml==0.7.1
fonttools==4.40.0
fpdf2==2.7.4
pefile==2023.2.7
Pillow==9.5.0
psycopg2-binary==2.9.6
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.4
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQt5-stubs==5.15.6.0
pywin32-ctypes==0.2.2
